refactor: move getFeed.ts to users
